index function bug fixed

+++ Algorithms/radix sort/radix_sort.py
@@ -2,7 +2,7 @@
 
 
 def index(i: int, exp: int) -> int:
-    return (i/exp) % 10
+    return int((i/exp) % 10)
